+++ src/zip.h
@@ -1,4 +1,4 @@
-// This code was copied from the R package 'readxl' at 
+// This code was adapted from the R package 'readxl' at
 // https://cran.r-project.org/web/packages/readxl/index.html
 // on 4 July 2016.
 // It was written by Hadley Wickham hadley@rstudio.com.
@@ -12,6 +12,7 @@
 
 std::string zip_buffer(const std::string& zip_path, const std::string& file_path);
 bool zip_has_file(const std::string& zip_path, const std::string& file_path);
+std::string extdata();
 
 #endif